requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.4
psycopg2-binary
Pillow
python-dotenv==1.0.0 